Compiler: fix crash when suspend body contains only an exit.
A Suspend gate needs to connect to a K1 return code. In some
situations (like only an Exit in the body) there is no K1 wire. This
commit add a check to avoid the connection in that case.

The patch also update the Exit in order to avoid creating empty arrays
for K[x] where x < exit return code.